+++ Example/loudspeaker/LSPViewController.h
@@ -14,7 +14,7 @@
 @property (weak, nonatomic) IBOutlet UIButton *demoButton;
 @property (weak, nonatomic) IBOutlet UIButton *harlequinButton;
 
-- (void)closeDemo;
+- (void)resetDemoButtons;
 - (IBAction)launchDemo:(id)sender;
 - (IBAction)launchHarlequinDemo:(id)sender;