+++ Pod/Classes/LSPAudioPlayerModel.m
@@ -1,6 +1,6 @@
 //
 //  LSPAudioPlayerModel.m
-//  Expecta
+//  loudspeaker
 //
 //  Created by Adam Yanalunas on 11/16/17.
 //