+++ CHANGELOG.md
@@ -10,9 +10,10 @@ This release contains breaking changes so read the changes carefully to see how 
 * Added iPhone X/safe area layout support
 * Added icons and loading screen to example project
 * Added configuration
-  * Allows configuration of player height, volume, interval to update progress bar and elapsed string, and accuracy for taps on progress to jump playback
+  * Allows configuration of volume, interval to update progress bar and elapsed string, and accuracy for taps on progress to jump playback
   * All `LSPAudioViewController` instances must be initialized with a `LSPConfiguration` instance
   * A default configuration is available from `deafultConfiguration` in `LSPConfigurationBuilder` or you can build your own
+  * Player frame is now determined by `LSPAudioViewController`'s frame
 * Added `show` and `hide:` to `LSPAudioViewController` to toggle the default presentation animations
 * Added accessibility to interactive elements
   * Accessibility elements are localized to English, localization strings available to be overridden by other languages